requirements.txt
streamlit 
pillow